# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * @nullable
 */
export type InvitationInputRole = typeof InvitationInputRole[keyof typeof InvitationInputRole] | null;


export const InvitationInputRole = {
  builder: 'builder',
  investor: 'investor',
  founder: 'founder',
  company: 'company',
  /** @deprecated use builder */
  operator: 'operator',
} as const;

export interface InvitationInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  role?: InvitationInputRole;
  /** @nullable */
  linkedin?: string | null;
  /** @minLength 1 */
  whoYouAre: string;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  whoIntroduced?: string | null;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  organizationDomain?: string | null;
  /** @nullable */
  organizationLogo?: string | null;
  /**
     * Honeypot field — must be empty
     * @nullable
     */
  fax?: string | null;
  /**
     * Legacy honeypot field — must be empty
     * @nullable
     */
  company?: string | null;
}

export interface InvitationResponse {
  message: string;
}

export interface InvitationRecord {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  linkedin?: string | null;
  whoYouAre: string;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  whoIntroduced?: string | null;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  organizationDomain?: string | null;
  /** @nullable */
  organizationLogo?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListRequestsParams = {
passcode: string;
};