# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: invitation
    description: Invitation request operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /request:
    post:
      operationId: submitRequest
      tags: [invitation]
      summary: Submit an invitation request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationInput"
      responses:
        "201":
          description: Request received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /requests:
    get:
      operationId: listRequests
      tags: [admin]
      summary: List all invitation requests (admin only)
      parameters:
        - name: passcode
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of invitation requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvitationRecord"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    InvitationInput:
      type: object
      required:
        - name
        - email
        - whoYouAre
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        role:
          type: ["string", "null"]
          enum: ["builder", "investor", "founder", "company", "operator", null]
        linkedin:
          type: ["string", "null"]
        whoYouAre:
          type: string
          minLength: 1
        link:
          type: ["string", "null"]
        whoIntroduced:
          type: ["string", "null"]
        organization:
          type: ["string", "null"]
          description: Company / fund / firm name (e.g. Sequoia, a16z)
        organizationDomain:
          type: ["string", "null"]
          description: Company website domain used to resolve logo (e.g. sequoiacap.com)
        organizationLogo:
          type: ["string", "null"]
          description: Optional pre-resolved logo URL; server may override after caching
        fax:
          type: ["string", "null"]
          description: Honeypot field — must be empty
        company:
          type: ["string", "null"]
          description: Legacy honeypot field — must be empty
    InvitationResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    InvitationRecord:
      type: object
      required:
        - id
        - name
        - email
        - whoYouAre
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: ["string", "null"]
        linkedin:
          type: ["string", "null"]
        whoYouAre:
          type: string
        link:
          type: ["string", "null"]
        whoIntroduced:
          type: ["string", "null"]
        organization:
          type: ["string", "null"]
        organizationDomain:
          type: ["string", "null"]
        organizationLogo:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string