# lib/api-zod/src/generated/types/invitationInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvitationInputRole } from './invitationInputRole';

export interface InvitationInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  role?: InvitationInputRole;
  /** @nullable */
  linkedin?: string | null;
  /** @minLength 1 */
  whoYouAre: string;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  whoIntroduced?: string | null;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  organizationDomain?: string | null;
  /** @nullable */
  organizationLogo?: string | null;
  /**
     * Honeypot field — must be empty
     * @nullable
     */
  fax?: string | null;
  /**
     * Legacy honeypot field — must be empty
     * @nullable
     */
  company?: string | null;
}