# lib/api-zod/src/generated/types/invitationRecord.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InvitationRecord {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  linkedin?: string | null;
  whoYouAre: string;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  whoIntroduced?: string | null;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  organizationDomain?: string | null;
  /** @nullable */
  organizationLogo?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  createdAt: Date;
}